Adds lgtm changes and addresses core dev comments

+++ psi4/src/psi4/lib3index/cholesky.cc
@@ -163,7 +163,7 @@ CholeskyERI::CholeskyERI(std::shared_ptr<TwoBodyAOInt> integral, double schwarz,
     basisset_ = integral_->basis();
 }
 CholeskyERI::~CholeskyERI() {}
-size_t CholeskyERI::N() { return basisset_->nbf() * basisset_->nbf(); }
+size_t CholeskyERI::N() { return static_cast<size_t>(basisset_->nbf()) * basisset_->nbf(); }
 void CholeskyERI::compute_diagonal(double* target) {
     const double* buffer = integral_->buffer();
     for (size_t M = 0; M < basisset_->nshell(); M++) {
@@ -265,7 +265,7 @@ CholeskyDelta::CholeskyDelta(std::shared_ptr<Vector> eps_aocc, std::shared_ptr<V
                              size_t memory)
     : eps_aocc_(eps_aocc), eps_avir_(eps_avir), Cholesky(delta, memory) {}
 CholeskyDelta::~CholeskyDelta() {}
-size_t CholeskyDelta::N() { return eps_aocc_->dimpi()[0] * eps_avir_->dimpi()[0]; }
+size_t CholeskyDelta::N() { return static_cast<size_t>(eps_aocc_->dimpi()[0]) * eps_avir_->dimpi()[0]; }
 void CholeskyDelta::compute_diagonal(double* target) {
     size_t naocc = eps_aocc_->dimpi()[0];
     size_t navir = eps_avir_->dimpi()[0];
